Move static field set to top level class
Also split into two classes, consistent with how arrays are done
